% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLFM.R
\name{SLFM}
\alias{SLFM}
\title{SLFM1D}
\usage{
SLFM(
  ymat,
  n_pos = 1000,
  burnin = 500,
  thin = 1,
  hyperparams = list(a = 1, b = 1/10, ccc_a = 1, ccc_b = 25, kappa_a = 1, omega_sd =
    0.1, kappa_sd = 0.5, j_epi = 0.04, i_epi = 0.02, j_leap = 5, i_leap = 5, skip = 50,
    jitter = T),
  initial_values = NULL,
  core = 2,
  cluster_seed = 1234
)
}
\arguments{
\item{ymat}{A binary matrix containing 0 and 1 or missing values(NA)}

\item{n_pos}{Number of posterior samplers to collect}

\item{burnin}{Burnin/warmup period for the MCMC}

\item{thin}{Thining of the posterior chain, the default is 1}

\item{hyperparams}{Hyperparameters}

\item{initial_values}{Initial values for input in the order of \eqn{kappa_j}, $beta_i$, $zeta_j$, $psi_j$,$omega$,$xi_inv$ (inverse of xi)}

\item{core}{Number of cpu cores to use for parallel computing}

\item{cluster_seed}{Seed number for reproducibility when 2 or more cores are used.}
}
\value{
A list containing the posterior samples for each model parameter
}
\description{
Fits a one-dimensional Spherical Latent Factor Model (Circular Factor Model) on a dataset consists of binary responses
}
\examples{
data(ymat)
set.seed(2021)
posterior_chain = SLFM(ymat = ymat,n_pos=200,burnin=50,core=2)
}
